Replace macro with base type
This loses type the type safety of ensuring we init the struct with
the correct port pin, but makes the code much easier to write & reason
about.

I'm happy to make this trade off since I'm providing the init method and
can manually verify the one instantiation of the leds.